requirements.txt
tree_sitter
GitPython
loguru
sexpdata

# Development
ipython
